requirements.txt
Django~=1.10
django-ckeditor==5.3.0
Pillow==5.3.0